Exception handling - delete and update

+++ src/main/java/com/ProjetoWebServices/ProjetoWebServices/services/UserService.java
@@ -1,13 +1,16 @@
 package com.ProjetoWebServices.ProjetoWebServices.services;
 
-import java.util.List;
+import java.util.List; 
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.ProjetoWebServices.ProjetoWebServices.entities.User;
 import com.ProjetoWebServices.ProjetoWebServices.repositories.UserRepository;
+import com.ProjetoWebServices.ProjetoWebServices.services.exceptions.DatabaseException;
 import com.ProjetoWebServices.ProjetoWebServices.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -27,12 +30,27 @@ public class UserService {
 		return repository.save(obj);
 	}
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try{
+			if(repository.existsById(id)){
+				repository.deleteById(id);
+			}
+			else {
+				throw new ResourceNotFoundException(id);
+			}
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	public User update(Long id, User obj) {
-		User entity = repository.getReferenceById(id); //não vai no banco de dados, ele retorna um objeto monitorado pelo jpa
-		updateData(entity, obj);
-		return repository.save(entity);
+		if(repository.existsById(id)){
+			User entity = repository.getReferenceById(id); //não vai no banco de dados, ele retorna um objeto monitorado pelo jpa
+			updateData(entity, obj);
+			return repository.save(entity);
+		}
+		else {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
